requirements.txt
altair==6.0.0
APScheduler==3.11.2
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.3
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.16.0
numpy==2.4.2
packaging==26.0
pandas==2.3.3
pillow==12.1.1
playwright==1.58.0
playwright-stealth==2.0.2
protobuf==6.33.5
pyarrow==23.0.1
pydeck==0.9.1
pyee==13.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
smmap==5.0.2
soupsieve==2.8.3
streamlit==1.54.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
watchdog==6.0.0
Werkzeug==3.1.6
psycopg2-binary
python-dotenv
supabase
duffel-api